requirements.txt
# ====================================
# Credit Risk Scoring Project
# Python Dependencies
# ====================================

# -----------------
# Core Data Science
# -----------------
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# -----------------
# Machine Learning
# -----------------
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
imbalanced-learn>=0.11.0

# -----------------
# Hyperparameter Tuning
# -----------------
optuna>=3.4.0

# -----------------
# Model Explainability
# -----------------
shap>=0.43.0

# -----------------
# Database
# -----------------
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0

# -----------------
# API
# -----------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# -----------------
# UI
# -----------------
streamlit>=1.28.0

# -----------------
# Orchestration
# -----------------
apache-airflow>=2.7.0

# -----------------
# Monitoring
# -----------------
prometheus-client>=0.18.0

# -----------------
# Visualization
# -----------------
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# -----------------
# Configuration & Utils
# -----------------
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
tqdm>=4.66.0
joblib>=1.3.0

# -----------------
# Data Validation
# -----------------
great-expectations>=0.18.0

# -----------------
# Testing
# -----------------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# -----------------
# Code Quality
# -----------------
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0

# -----------------
# Notebooks
# -----------------
jupyter>=1.0.0
ipykernel>=6.26.0
nbformat>=5.9.0

# -----------------
# Kaggle
# -----------------
kaggle>=1.5.16
